backend/requirements.txt
asgiref
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv